Co-locating page files to make discovery and refactoring easier.
