+++ todoodlefrontend/src/pages/TodoPage/TodoPage.tsx
@@ -1,12 +1,12 @@
-import React, { useState, useEffect } from "react"
-import { useAuth } from "../hooks/useAuth"
-import NavBar from "../components/NavBar/NavBar"
-import TodoForm from "../components/Todo/TodoForm"
-import TaskModal from "../components/Todo/TaskModal"
-import type { Task, Status, Folder } from "../components/Todo/types"
-import { notesApi, foldersApi } from "../services/api"
-import { noteToTask, taskToNote, addDefaultColors } from "../utils/dataTransformers"
-import "./Page.css"
+import { useState, useEffect } from "react"
+import { useAuth } from "../../hooks/useAuth"
+import NavBar from "../../components/NavBar/NavBar"
+import TodoForm from "../../components/Todo/TodoForm"
+import TaskModal from "../../components/Todo/TaskModal"
+import type { Task, Status, Folder } from "../../components/Todo/types"
+import { notesApi, foldersApi } from "../../services/api"
+import { noteToTask, taskToNote, addDefaultColors } from "../../utils/dataTransformers"
+import "../Page.css"
 
 type SortOption = 'title' | 'status' | 'folder' | 'createdAt' | 'priority';
 type SortDirection = 'asc' | 'desc';
