+++ todoodlefrontend/src/pages/UserSignUpPage/UserSignUpPage.tsx
@@ -1,9 +1,10 @@
 import React from "react"
 import { useState } from "react"
 import { useNavigate, Link } from "react-router-dom"
-import "../styles/globals.css"
-import NavBar from "../components/NavBar/NavBar"
-import logo from "../assets/virtual-learning-background-with-design-space.png"
+import "../../styles/globals.css"
+import "../Page.css"
+import NavBar from "../../components/NavBar/NavBar"
+import logo from "../../assets/virtual-learning-background-with-design-space.png"
 
 const UserSignUpPage = () => {
   const navigate = useNavigate()
